move some states to their own files
