+++ src/components/Spinner/index.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {render, screen} from '@testing-library/react';
-import {Spinner} from '..';
+import {Spinner} from '.';
 
 describe('Spinner', () => {
     it('should render spinner', () => {
